feat: new module stellar-base
