+++ types/stellar-sdk/index.d.ts
@@ -11,464 +11,40 @@
 
 /// <reference types="node" />
 
-// Dependency
-export type Key = string|number|symbol;
-export type Diff<T extends Key, U extends Key> = ({[P in T]: P } & {[P in U]: never } & { [x: string]: never })[T];
-export type Omit<T, K extends keyof T> = Pick<T, Diff<keyof T, K>>;
-
-export namespace StellarBase {
-    class Account {
-        constructor(accountId: string, sequence: string | number)
-        accountId(): string;
-        sequenceNumber(): string;
-        incrementSequenceNumber(): void;
-    }
-
-    enum ASSET_TYPE {
-        native = 'native',
-        credit4 = 'credit_alphanum4',
-        credit12 = 'credit_alphanum12',
-    }
-    class Asset {
-        static native(): Asset;
-        static fromOperation(xdr: xdr.Asset): Asset;
-
-        constructor(code: string, issuer: string);
-
-        getCode(): string;
-        getIssuer(): string;
-        getAssetType(): ASSET_TYPE;
-        isNative(): boolean;
-        equals(other: Asset): boolean;
-        toXDRObject(): xdr.Asset;
-
-        code: string;
-        issuer: string;
-    }
-
-    const FastSigning: boolean;
-
-    class Keypair {
-        static fromRawEd25519Seed(secretSeed: Buffer): Keypair;
-        static fromBase58Seed(secretSeed: string): Keypair;
-        static fromSecret(secretKey: string): Keypair;
-        static master(): Keypair;
-        static fromPublicKey(publicKey: string): Keypair;
-        static random(): Keypair;
-
-        constructor(keys: { type: 'ed25519', secretKey: string } | { type: 'ed25519', Key: string })
-
-        publicKey(): string;
-        secret(): string;
-        rawPublicKey(): Buffer;
-        rawSecretKey(): Buffer;
-        canSign(): boolean;
-        sign(data: Buffer): Buffer;
-        signDecorated(data: Buffer): xdr.DecoratedSignature;
-        signatureHint(): xdr.SignatureHint;
-        verify(data: Buffer, signature: Buffer): boolean;
-    }
-
-    const MemoNone = 'none';
-    const MemoID = 'id';
-    const MemoText = 'text';
-    const MemoHash = 'hash';
-    const MemoReturn = 'return';
-    type MemoNone = 'none';
-    type MemoID = 'id';
-    type MemoText = 'text';
-    type MemoHash = 'hash';
-    type MemoReturn = 'return';
-    class Memo<T extends Memo.AnyType = Memo.AnyType> {
-        static fromXDRObject(memo: xdr.Memo): Memo;
-        static hash(hash: string): Memo<MemoHash>;
-        static id(id: string): Memo<MemoID>;
-        static none(): Memo<MemoNone>;
-        static return(hash: string): Memo<MemoReturn>;
-        static text(text: string): Memo<MemoText>;
-
-        constructor(type: MemoNone, value?: null);
-        constructor(type: MemoHash | MemoReturn, value: Buffer)
-        constructor(type: MemoHash | MemoReturn | MemoID | MemoText, value: string)
-        constructor(type: T, value: Memo.AnyValue)
-
-        type: T;
-        value:
-            T extends MemoNone ? null :
-            T extends MemoID ? string :
-            T extends MemoText ? string :
-            T extends MemoHash ? Buffer :
-            T extends MemoReturn ? Buffer :
-            Memo.AnyValue;
-
-        toXDRObject(): xdr.Memo;
-    }
-    namespace Memo {
-        type AnyType = MemoNone | MemoID | MemoText | MemoHash | MemoReturn;
-        type AnyValue = null | string | Buffer;
-    }
-
-    enum Networks {
-        PUBLIC = 'Public Global Stellar Network ; September 2015',
-        TESTNET = 'Test SDF Network ; September 2015',
-    }
-
-    class Network {
-        static use(network: Network): void;
-        static usePublicNetwork(): void;
-        static useTestNetwork(): void;
-        static current(): Network;
-
-        constructor(passphrase: string)
-
-        networkPassphrase(): string;
-        networkId(): string;
-    }
-
-    const AuthRequiredFlag: 1;
-    const AuthRevocableFlag: 2;
-    const AuthImmutableFlag: 4;
-
-    type TransactionOperation =
-        Operation.CreateAccount
-        | Operation.Payment
-        | Operation.PathPayment
-        | Operation.CreatePassiveOffer
-        | Operation.ManageOffer
-        | Operation.SetOptions
-        | Operation.ChangeTrust
-        | Operation.AllowTrust
-        | Operation.AccountMerge
-        | Operation.Inflation
-        | Operation.ManageData
-        | Operation.BumpSequence;
-
-    enum OperationType {
-        createAccount = 'createAccount',
-        payment = 'payment',
-        pathPayment = 'pathPayment',
-        createPassiveOffer = 'createPassiveOffer',
-        manageOffer = 'manageOffer',
-        setOptions = 'setOptions',
-        changeTrust = 'changeTrust',
-        allowTrust = 'allowTrust',
-        accountMerge = 'accountMerge',
-        inflation = 'inflation',
-        manageData = 'manageData',
-        bumpSequence = 'bumpSequence',
-    }
-
-    namespace Operation {
-        interface Operation<T extends OperationType = OperationType> {
-            type: T;
-            source?: string;
-        }
-        interface OperationOptions {
-            source?: string;
-        }
-
-        interface AccountMerge extends Operation<OperationType.accountMerge> {
-            destination: string;
-        }
-        interface AccountMergeOptions extends OperationOptions {
-            destination: string;
-        }
-        function accountMerge(options: AccountMergeOptions): xdr.Operation<AccountMerge>;
-
-        interface AllowTrust extends Operation<OperationType.allowTrust> {
-            trustor: string;
-            assetCode: string;
-            authorize: boolean | undefined;
-        }
-        interface AllowTrustOptions extends OperationOptions {
-            trustor: string;
-            assetCode: string;
-            authorize?: boolean;
-        }
-        function allowTrust(options: AllowTrustOptions): xdr.Operation<AllowTrust>;
-
-        interface ChangeTrust extends Operation<OperationType.changeTrust> {
-            line: Asset;
-            limit: string;
-        }
-        interface ChangeTrustOptions extends OperationOptions {
-            asset: Asset;
-            limit?: string;
-        }
-        function changeTrust(options: ChangeTrustOptions): xdr.Operation<ChangeTrust>;
-
-        interface CreateAccount extends Operation<OperationType.createAccount> {
-            destination: string;
-            startingBalance: string;
-        }
-        interface CreateAccountOptions extends OperationOptions {
-            destination: string;
-            startingBalance: string;
-        }
-        function createAccount(options: CreateAccountOptions): xdr.Operation<CreateAccount>;
-
-        interface CreatePassiveOffer extends Operation<OperationType.createPassiveOffer> {
-            selling: Asset;
-            buying: Asset;
-            amount: string;
-            price: string;
-        }
-        interface CreatePassiveOfferOptions extends OperationOptions {
-            selling: Asset;
-            buying: Asset;
-            amount: string;
-            price: number | string | object /* bignumber.js */;
-        }
-        function createPassiveOffer(options: CreatePassiveOfferOptions): xdr.Operation<CreatePassiveOffer>;
-
-        interface Inflation extends Operation<OperationType.inflation> {
-        }
-        interface InflationOptions extends OperationOptions {  // tslint:disable-line
-        }
-        function inflation(options: InflationOptions): xdr.Operation<Inflation>;
-
-        interface ManageData extends Operation<OperationType.manageData> {
-            name: string;
-            value: Buffer;
-        }
-        interface ManageDataOptions extends OperationOptions {
-            name: string;
-            value: string | Buffer;
-        }
-        function manageData(options: ManageDataOptions): xdr.Operation<ManageData>;
-
-        interface ManageOffer extends Operation<OperationType.manageOffer> {
-            selling: Asset;
-            buying: Asset;
-            amount: string;
-            price: string;
-            offerId: string;
-        }
-        interface ManageOfferOptions extends CreatePassiveOfferOptions {
-            offerId?: number | string;
-        }
-        function manageOffer(options: ManageOfferOptions): xdr.Operation<ManageOffer>;
-
-        interface PathPayment extends Operation<OperationType.pathPayment> {
-            sendAsset: Asset;
-            sendMax: string;
-            destination: string;
-            destAsset: Asset;
-            destAmount: string;
-            path: Asset[];
-        }
-        interface PathPaymentOptions extends OperationOptions {
-            sendAsset: Asset;
-            sendMax: string;
-            destination: string;
-            destAsset: Asset;
-            destAmount: string;
-            path?: Asset[];
-        }
-        function pathPayment(options: PathPaymentOptions): xdr.Operation<PathPayment>;
-
-        interface Payment extends Operation<OperationType.payment> {
-            amount: string;
-            asset: Asset;
-            destination: string;
-        }
-        interface PaymentOptions extends OperationOptions {
-            amount: string;
-            asset: Asset;
-            destination: string;
-        }
-        function payment(options: PaymentOptions): xdr.Operation<Payment>;
-
-        /*
-        * Required = 1 << 0
-        * Revocable = 1 << 1
-        * Immutable = 1 << 2
-        */
-        enum AuthFlags {
-            Required = 1,
-            Revocable = 2,
-            Immutable = 4,
-        }
-        interface SignerEd25519PublicKey {
-            ed25519PublicKey: string;
-            weight: number | undefined;
-        }
-        interface SignerSha256Hash {
-            sha256Hash: Buffer;
-            weight: number | undefined;
-        }
-        interface SignerPreAuthTx {
-            preAuthTx: Buffer;
-            weight: number | undefined;
-        }
-        type Signer = SignerEd25519PublicKey | SignerSha256Hash | SignerPreAuthTx;
-        interface SignerEd25519PublicKeyOptions {
-            ed25519PublicKey: string;
-            weight?: number | string;
-        }
-        interface SignerSha256HashOptions {
-            sha256Hash: Buffer | string;
-            weight?: number | string;
-        }
-        interface SignerPreAuthTxOptions {
-            preAuthTx: Buffer | string;
-            weight?: number | string;
-        }
-        type SignerOptions = SignerEd25519PublicKeyOptions | SignerSha256HashOptions | SignerPreAuthTxOptions;
-        type SignerUnion = {ed25519PublicKey: any} | {sha256Hash: any} | {preAuthTx: any} | null;
-        interface SetOptions<T extends SignerUnion = SignerUnion> extends Operation<OperationType.setOptions> {
-            inflationDest?: string;
-            clearFlags?: AuthFlags;
-            setFlags?: AuthFlags;
-            masterWeight?: number;
-            lowThreshold?: number;
-            medThreshold?: number;
-            highThreshold?: number;
-            homeDomain?: string;
-            signer:
-                T extends {ed25519PublicKey: any} ? SignerEd25519PublicKey :
-                T extends {sha256Hash: any} ? SignerSha256Hash :
-                T extends {preAuthTx: any} ? SignerPreAuthTx :
-                never;
-        }
-        interface SetOptionsOptions<T extends SignerUnion = never> extends OperationOptions {
-            inflationDest?: string;
-            clearFlags?: AuthFlags;
-            setFlags?: AuthFlags;
-            masterWeight?: number | string;
-            lowThreshold?: number | string;
-            medThreshold?: number | string;
-            highThreshold?: number | string;
-            homeDomain?: string;
-            signer?: T;
-        }
-        function setOptions<T extends SignerUnion = never>(options: SetOptionsOptions<T>): xdr.Operation<SetOptions<T>>;
-
-        interface BumpSequence extends Operation<OperationType.bumpSequence> {
-            bumpTo: string;
-        }
-        interface BumpSequenceOptions extends OperationOptions {
-            bumpTo: string;
-        }
-        function bumpSequence(options: BumpSequenceOptions): xdr.Operation<BumpSequence>;
-
-        function fromXDRObject<T extends TransactionOperation = TransactionOperation>(xdrOperation: xdr.Operation<T>): T;
-    }
-
-    namespace StrKey {
-        function encodeEd25519PublicKey(data: Buffer): string;
-        function decodeEd25519PublicKey(data: string): Buffer;
-        function isValidEd25519PublicKey(Key: string): boolean;
-
-        function encodeEd25519SecretSeed(data: Buffer): string;
-        function decodeEd25519SecretSeed(data: string): Buffer;
-        function isValidEd25519SecretSeed(seed: string): boolean;
-
-        function encodePreAuthTx(data: Buffer): string;
-        function decodePreAuthTx(data: string): Buffer;
-
-        function encodeSha256Hash(data: Buffer): string;
-        function decodeSha256Hash(data: string): Buffer;
-    }
-
-    class Transaction<TMemo extends Memo = Memo, TOps extends TransactionOperation[] = TransactionOperation[]> {
-        constructor(envelope: string | xdr.TransactionEnvelope)
-        hash(): Buffer;
-        sign(...keypairs: Keypair[]): void;
-        signatureBase(): Buffer;
-        signHashX(preimage: Buffer | string): void;
-        toEnvelope(): xdr.TransactionEnvelope;
-
-        operations: TOps;
-        sequence: number;
-        fee: number;
-        source: string;
-        memo: TMemo;
-        signatures: xdr.DecoratedSignature[];
-    }
-
-    class TransactionBuilder {
-        constructor(sourceAccount: Account, options?: TransactionBuilder.TransactionBuilderOptions)
-        addOperation(operation: xdr.Operation): this;
-        addMemo(memo: Memo): this;
-        build(): Transaction;
-    }
-
-    namespace TransactionBuilder {
-        interface TransactionBuilderOptions {
-            fee?: number;
-            timebounds?: {
-                minTime?: number | string
-                maxTime?: number | string
-            };
-            memo?: Memo;
-        }
-    }
-
-    namespace xdr {
-        class XDRStruct {
-            static fromXDR(xdr: Buffer): XDRStruct;
-
-            toXDR(base?: string): Buffer;
-            toXDR(encoding: string): string;
-        }
-        class Operation<T extends TransactionOperation = TransactionOperation> extends XDRStruct {
-            static fromXDR(xdr: Buffer): Operation;
-        }
-        class Asset extends XDRStruct {
-            static fromXDR(xdr: Buffer): Asset;
-        }
-        class Memo extends XDRStruct {
-            static fromXDR(xdr: Buffer): Memo;
-        }
-        class TransactionEnvelope extends XDRStruct {
-            static fromXDR(xdr: Buffer): TransactionEnvelope;
-        }
-        class DecoratedSignature extends XDRStruct {
-            static fromXDR(xdr: Buffer): DecoratedSignature;
-
-            constructor(keys: { hint: SignatureHint, signature: Signature })
-
-            hint(): SignatureHint;
-            signature(): Buffer;
-        }
-        type SignatureHint = Buffer;
-        type Signature = Buffer;
-
-        class TransactionResult extends XDRStruct {
-            static fromXDR(xdr: Buffer): TransactionResult;
-        }
-    }
-
-    function hash(data: Buffer): Buffer;
-    function sign(data: Buffer, rawSecret: Buffer): Buffer;
-    function verify(data: Buffer, signature: Buffer, rawPublicKey: Buffer): boolean;
-}
+import { ASSET_TYPE, Asset, Memo, Transaction } from 'stellar-base';
 
 // Re-StellarBase
-export import Account = StellarBase.Account;
-export import ASSET_TYPE = StellarBase.ASSET_TYPE;
-export import Asset = StellarBase.Asset;
-export import FastSigning = StellarBase.FastSigning;
-export import Keypair = StellarBase.Keypair;
-export import Memo = StellarBase.Memo;
-export import MemoHash = StellarBase.MemoHash;
-export import MemoID = StellarBase.MemoID;
-export import MemoNone = StellarBase.MemoNone;
-export import MemoReturn = StellarBase.MemoReturn;
-export import MemoText = StellarBase.MemoText;
-export import Network = StellarBase.Network;
-export import Networks = StellarBase.Networks;
-export import Operation = StellarBase.Operation;
-export import OperationType = StellarBase.OperationType;
-export import StrKey = StellarBase.StrKey;
-export import Transaction = StellarBase.Transaction;
-export import TransactionBuilder = StellarBase.TransactionBuilder;
-export import TransactionOperation = StellarBase.TransactionOperation;
-export import hash = StellarBase.hash;
-export import sign = StellarBase.sign;
-export import verify = StellarBase.verify;
-export import xdr = StellarBase.xdr;
+export {
+    Account,
+    ASSET_TYPE,
+    Asset,
+    FastSigning,
+    Keypair,
+    Memo,
+    MemoHash,
+    MemoID,
+    MemoNone,
+    MemoReturn,
+    MemoText,
+    Network,
+    Networks,
+    Operation,
+    OperationType,
+    StrKey,
+    Transaction,
+    TransactionBuilder,
+    TransactionOperation,
+    hash,
+    sign,
+    verify,
+    xdr
+} from 'stellar-base';
+
+// Shorthands, not-to-export.
+export {};
+type Key = string|number|symbol;
+type Diff<T extends Key, U extends Key> = ({[P in T]: P } & {[P in U]: never } & { [x: string]: never })[T];
+type Omit<T, K extends keyof T> = Pick<T, Diff<keyof T, K>>;
 
 export class NetworkError extends Error {
     private response: any;
